+++ Editor/Drawers/Decorator/SepTitleAttributeDrawer.cs
@@ -1,5 +1,4 @@
-﻿using SaintsField.Utils;
-using UnityEditor;
+﻿using UnityEditor;
 using UnityEngine;
 
 namespace SaintsField.Editor.Drawers.Decorator